scripts/requirements.txt
beautifulsoup4==4.10.0
markdown==3.3.4
python-decouple==3.5
requests==2.26.0
